+++ CHANGELOG.md
@@ -10,7 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 - Add `statusContext` property to `<ProposalDataListItem />` module component
-- Add 'ADVANCEABLE' status for Proposals and ProposalVoting
+- Update `<ProposalDataListItem />` module component to support `ADVANCEABLE` status
 
 ## Changed